Almost done with DBUtils.java
I have commented everything thoroughly in the DBUtils.java file.

Removed all of the unnecessary methods and fixed most of the warnings.

Made a second table in DB with the name of debitCards in which table all of the debit cards are going to be stored there.

Finished InsertDebitCard() and CheckCardInDB(). Also did some bug fixing,